Backend/requirements.txt
django
djangorestframework
psycopg2-binary
django-filter
django-cors-headers
python-decouple
requests
bs4

